Move `NoStreamAccessAlert` into separate component.
